Uroboro.Tree

Description

The program representations we work on.

Documentation

type Identifier = String

This is used for type names, function names, constructor and destructor names, as well as variable names.

newtype Type

Represents both positive and negative data types.

Constructors

Type Identifier

data PExp

Parse Tree

Expression (Term).

Constructors

PVar Location Identifier
Variable.

PApp Location Identifier [PExp]
Constructor or function application.

PDes Location Identifier [PExp] PExp
Destructor application (Selection).

data PP

Pattern.

Constructors

PPVar Location Identifier
Variable pattern.

PPCon Location Identifier [PP]
Constructor pattern.

data PQ

Copattern.

Constructors

PQApp Location Identifier [PP]
Hole pattern.

PQDes Location Identifier [PP] PQ
Destructor pattern.

data PTCon

Constructor definition.

Constructors

PTCon Location Type Identifier [Type]

data PTDes

Destructor definition. Return type first, type to destruct last.

Constructors

PTDes Location Type Identifier [Type] Type

data PTRule

Part of a function definition.

Constructors

PTRule Location PQ PExp

data PT

Definition.

Constructors

PTPos Location Type [PTCon]
Data type.

PTNeg Location Type [PTDes]
Codata type.

PTFun Location Identifier [Type] Type [PTRule]
Function.

data TExp

Typed Syntax Tree

Expression with type annotations.

Constructors

TVar Type Identifier
Variable.

TApp Type Identifier [TExp]
Function application.

TCon Type Identifier [TExp]
Constructor application.

TDes Type Identifier [TExp] TExp
Destructor application.

data TP

Pattern with type annotations.

Constructors

TPVar Type Identifier

TPCon Type Identifier [TP]

data TQ

Copattern with type annotations.

Constructors

TQApp Type Identifier [TP]

TQDes Type Identifier [TP] TQ

type Rule = (TQ, TExp)

One rule of a function definition.

type Rules = [(Identifier, [Rule])]

A complete program.